fastapi_llm/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
python-multipart

langchain-core
langchain-ollama
langchain-chroma
langchain-huggingface
chromadb

datasets
pandas
sentence-transformers
transformers
sentencepiece

vosk
pydub
librosa
soundfile
numpy
scipy
gTTS
